Change telem_ctrl to talk directly with ArduPilot
telem_ctrl was made to talk to another process that will take the data,
process the additional header and pass the payload (that is a mavlink
packet) to the process/board running ArduPilot.

When we are using ArduPilot on Linux it's easier to just talk directly
with ArduPilot, so something like below will just work, also relaying
mavlink packets to the GCS:

	arducopter -A udp:10.1.1.1:14550

This is not intended for Solo as it will break their functionality.